requirements.txt
asgiref==3.5.2
certifi==2022.6.15
chardet==3.0.4
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pytz==2022.2.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
